fix: Refactor and unify error handling, token logic, axios configuration, … …and environment setup
