+++ frontend/src/config.ts
@@ -1,0 +1,3 @@
+export const config = {
+    apiUrl: process.env.REACT_APP_API_URL,
+};
