+++ frontend/src/pages/Login/Login.tsx
@@ -34,12 +34,10 @@ const Login = () => {
             toast.success(MESSAGES.SUCCES.USER_REGISTERED);
             setTimeout(() => navigate('/main-page'), 2000);
         } catch (error: unknown) {
-            if (typeof error === 'object' && error !== null && 'message' in error) {
-                toast.error((error as { message: string }).message);
-            } else if (error instanceof Error) {
-                toast.error(MESSAGES.ERROR.LOGIN_FAILED);
+            if (error instanceof Error) {
+                toast.error(error.message);
             } else {
-                toast.error('Coś poszło nie tak.');
+                toast.error(MESSAGES.ERROR.UNKNOWN);
             }
         }
     };
