+++ frontend/src/utils/messages.ts
@@ -8,7 +8,8 @@ export const MESSAGES = {
         INVALID_EMAIL: 'Nieprawidłowy email.',
         INVALID_PASSWORD: 'Hasło musi mieć co najmniej 8 znaków, w tym: dużą literę, małą literę, cyfrę i znak specjalny.',
         NOT_MATCHES_PASSWORD: 'Hasła muszą być identyczne',
-        LOGIN_FAILED: 'Ups! Coś się nie zgadza. Sprawdź login i hasło.',
-        REGISTER_FAILED: 'Ups! Mamy problem. Spróbuj ponownie.'
+        LOGIN_FAILED: 'Ups! Mamy problem z logowaniem. Sprawdź dane i spróbuj ponownie.',
+        REGISTER_FAILED: 'Ups! Mamy problem z rejestracją. Sprawdź dane i spróbuj ponownie.',
+        UNKNOWN: 'Nieznany błąd. Spróbuj ponownie.'
     }
 };
